+++ tw-tasks-core/src/main/java/com/transferwise/tasks/processing/ITasksProcessingService.java
@@ -21,4 +21,6 @@ public interface ITasksProcessingService {
   }
 
   void addTaskTriggeringFinishedListener(Consumer<TaskTriggering> consumer);
+
+  void startProcessing();
 }