+++ tw-tasks-core/src/test/java/com/transferwise/tasks/stucktasks/TasksResumerTest.java
@@ -1,5 +1,6 @@
 package com.transferwise.tasks.stucktasks;
 
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyLong;
@@ -21,12 +22,13 @@ import com.transferwise.tasks.handler.interfaces.ITaskHandler;
 import com.transferwise.tasks.handler.interfaces.ITaskHandlerRegistry;
 import com.transferwise.tasks.handler.interfaces.ITaskProcessingPolicy;
 import com.transferwise.tasks.handler.interfaces.ITaskProcessingPolicy.StuckTaskResolutionStrategy;
+import com.transferwise.tasks.handler.interfaces.StuckDetectionSource;
 import com.transferwise.tasks.helpers.IMeterHelper;
+import com.transferwise.tasks.stucktasks.TasksResumer.StuckTaskResolutionStats;
 import com.transferwise.tasks.triggering.ITasksExecutionTriggerer;
 import java.time.Clock;
 import java.time.Duration;
 import java.time.ZonedDateTime;
-import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.Stream;
 import org.apache.curator.framework.CuratorFramework;
 import org.junit.jupiter.api.BeforeEach;
@@ -83,9 +85,8 @@ class TasksResumerTest {
         .setType("TEST")
         .setStatus(TaskStatus.SUBMITTED.name())
         .setVersionId(new TaskVersionId(UuidUtils.generatePrefixCombUuid(), 0));
-    AtomicInteger numResumed = new AtomicInteger();
-
-    service.handleStuckTask(task, numResumed, null, null);
+    var stats = new StuckTaskResolutionStats();
+    service.handleStuckTask(task, StuckDetectionSource.CLUSTER_WIDE_STUCK_TASKS_DETECTOR, stats);
 
     verify(tasksExecutionTriggerer).trigger(argThat(
         baseTask -> 1 == baseTask.getVersion() && task.getVersionId().getId().equals(baseTask.getId())
@@ -95,6 +96,8 @@ class TasksResumerTest {
         task.getVersionId().getVersion(),
         now.plusMinutes(10)
     );
+
+    assertThat(stats.getResumed()).isEqualTo(1);
   }
 
   @ParameterizedTest(name = "handleStuckTask respects StuckTaskResolutionStrategy {3}")
@@ -109,21 +112,19 @@ class TasksResumerTest {
     lenient().when(taskDao.setStatus(any(), any(), anyLong())).thenReturn(true);
 
     ITaskProcessingPolicy processingPolicy = Mockito.mock(ITaskProcessingPolicy.class);
-    when(processingPolicy.getStuckTaskResolutionStrategy(any())).thenReturn(resolutionStrategy);
+    when(processingPolicy.getStuckTaskResolutionStrategy(any(), any())).thenReturn(resolutionStrategy);
     lenient().when(processingPolicy.getProcessingDeadline(any())).thenReturn(null);
 
     when(taskHandlerRegistry.getTaskHandler(any())).thenReturn(Mockito.mock(ITaskHandler.class));
     when(taskHandlerRegistry.getTaskHandler(null).getProcessingPolicy(any())).thenReturn(processingPolicy);
 
     // triggering tasks handling
-    AtomicInteger numResumed = new AtomicInteger();
-    AtomicInteger numError = new AtomicInteger();
-    AtomicInteger numFailed = new AtomicInteger();
+    StuckTaskResolutionStats stats = new StuckTaskResolutionStats();
     StuckTask task = new ITaskDao.StuckTask()
         .setType("TEST")
         .setStatus(TaskStatus.PROCESSING.name())
         .setVersionId(new TaskVersionId(UuidUtils.generatePrefixCombUuid(), 0));
-    service.handleStuckTask(task, numResumed, numError, numFailed);
+    service.handleStuckTask(task, StuckDetectionSource.CLUSTER_WIDE_STUCK_TASKS_DETECTOR, stats);
 
     verify(tasksExecutionTriggerer, times(resumed)).trigger(argThat(
         baseTask -> baseTask.getVersion() == 1 && baseTask.getId().equals(task.getVersionId().getId())
@@ -143,9 +144,9 @@ class TasksResumerTest {
         TaskStatus.FAILED,
         task.getVersionId().getVersion()
     );
-    assertEquals(resumed, numResumed.get());
-    assertEquals(error, numError.get());
-    assertEquals(failed, numFailed.get());
+    assertEquals(resumed, stats.getResumed());
+    assertEquals(error, stats.getError());
+    assertEquals(failed, stats.getFailed());
   }
 
   private static Stream<Arguments> resolutionCasesForHandleStuckTransfers() {